Added language selection for real list of supported languages fetched from the backend.
Set up state for multiple speakers, including the displaying of the appropriate text in the appropriate box. Note: Need to figure out how to multiplex in socket_io so both translations can occur over the same socket.

Flipped top part of the UI to be facing the person you are talking to.